refactor: Removed all unused files and created one unique index (just for now)
